requirements.txt
Django==4.1.7
django-ninja==0.21.0
requests==2.28.2